+++ src/main/kotlin/racingcar/dto/car/CarsDto.kt
@@ -1,0 +1,3 @@
+package racingcar.dto.car
+
+class CarsDto(_cars: List<CarDto>) : List<CarDto> by _cars
