+++ src/main/kotlin/racingcar/model/car/Cars.kt
@@ -1,8 +1,5 @@
 package racingcar.model.car
 
-import racingcar.dto.car.CarDto
-import racingcar.utils.mapper.toDto
-import racingcar.utils.mapper.toModel
 import racingcar.utils.random.RandomGenerator
 
 class Cars(_cars: List<Car>) : List<Car> by _cars {
@@ -28,16 +25,8 @@ class Cars(_cars: List<Car>) : List<Car> by _cars {
         return Winners(this.filter { it.position == winnerStandard.position })
     }
 
-    fun toDto(): CarsDto = CarsDto(
-        this.map { car -> car.toDto() }
-    )
-
     companion object {
         private const val DUPLICATED_CAR_NAME_ERROR_MESSAGE =
             "중복된 자동차 이름이 존재합니다."
     }
-}
-
-class CarsDto(_cars: List<CarDto>) : List<CarDto> by _cars {
-    fun toModel(): Cars = Cars(this.map { it.toModel() })
 }
